Revert "휴일이 없는 달 예외처리 필요"
This reverts commit d906f92b7064cccd16df84207b0b614bf3f40c82.
