+++ src/main/java/com/app/businessBridge/global/holidayapi/dto/HoliDayDto.java
@@ -1,8 +1,6 @@
 package com.app.businessBridge.global.holidayapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonSetter;
-import com.fasterxml.jackson.annotation.Nulls;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -10,16 +8,36 @@ import lombok.ToString;
 import java.util.List;
 
 
+
 @Getter
 @Setter
 @ToString
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class HoliDayDto {
+    private Response response;
 
-    private String totalCount;
+    @Getter
+    public static class Response {
+        private Body body;
+    }
 
-    private String locdate;
+    @Getter
+    public static class Body {
+        private Items items;
 
-    private List<String> locdate;
+        private String totalCount;
 
+    }
+
+    @Getter
+    public static class Items {
+        private List<Item> item;
+
+    }
+
+    @Getter
+    public static class Item {
+        private String locdate;
+
+    }
 }
